chatbot-backend/requirements.txt
# Core Framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
gunicorn>=21.0.0,<23.0.0

# Validation & Configuration
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0

# LLM Providers
google-genai>=0.3.0,<1.0.0
groq>=0.4.0,<1.0.0

# Database
google-cloud-firestore>=2.14.0,<3.0.0
google-auth>=2.25.0,<3.0.0

# Rate Limiting
slowapi>=0.1.9,<1.0.0